requirements.txt
numpy==1.21.2
Pillow==8.3.1
yapf==0.31.0
